requirements.txt
Flask==2.2.3
reportlab==3.6.12
pdfkit==1.0.0
Werkzeug==2.2.3
Jinja2==3.1.2
gunicorn==20.1.0
python-dotenv==1.0.0
Flask-SQLAlchemy==3.0.3
mysqlclient==2.1.1
cryptography>=38.0.0